+++ torch/csrc/jit/ir.cpp
@@ -6,6 +6,7 @@
 #include "torch/csrc/jit/constants.h"
 #include "torch/csrc/jit/assertions.h"
 #include "torch/csrc/jit/script/compiler.h"
+#include "torch/csrc/jit/passes/pretty_print.h"
 
 #include <iostream>
 #include <unordered_map>
@@ -72,88 +73,6 @@ std::ostream& operator<<(std::ostream & out, const_value_list_with_types l) {
   }
   return out;
 }
-template<typename T>
-void printPrimList(std::ostream & out, const std::vector<T> & items) {
-  out << "[";
-  int i = 0;
-  for(auto & item : items) {
-    if(i++ > 0)
-      out << ", ";
-    out << item;
-  }
-  out << "]";
-}
-
-std::string escapeString(std::string s) {
-  std::vector<char> search = {'\n', '\t', '\v'};
-  std::vector<std::string> replace = {"\\n", "\\t", "\\v"};
-  for (size_t i = 0; i < search.size(); i++) {
-    size_t pos = s.find(search[i]);
-    while(pos != std::string::npos) {
-      s.replace(pos, 1, replace[i]);
-      pos = s.find(search[i], pos + 1);
-    }
-  }
-  return s;
-}
-
-void printAttributeValue(std::ostream & out, Symbol & name, const Node * n) {
-  switch(n->kindOf(name)) {
-    case AttributeKind::f:
-      out << n->f(name);
-      break;
-    case AttributeKind::fs:
-      printPrimList(out,n->fs(name));
-      break;
-    case AttributeKind::i:
-      out << n->i(name);
-      break;
-    case AttributeKind::is:
-      printPrimList(out,n->is(name));
-      break;
-    case AttributeKind::s:
-      out << escapeString(n->s(name));
-      break;
-    case AttributeKind::ss:
-      printPrimList(out,n->ss(name));
-      break;
-    case AttributeKind::t:
-      {
-        at::Tensor t = n->t(name);
-        // 1-elem tensors are usually boxed scalars, so print them like it
-        if (t.numel() == 1) {
-          auto scalar_tensor = at::_local_scalar(t.view({}));
-          out << "{";
-          if (scalar_tensor.isFloatingPoint()) {
-            out << scalar_tensor.toDouble();
-          } else {
-            out << scalar_tensor.toLong();
-          }
-          out << "}";
-        } else if (t.numel() <= max_tensor_display_size) {
-          // TODO: This is awful code.  Also it doesn't work on Windows.
-          std::ostringstream tensor_ss;
-          tensor_ss << t;
-          std::string tensor_s{tensor_ss.str()};
-          // Remove newlines
-          std::replace(tensor_s.begin(), tensor_s.end(), '\n', ' ');
-          out << tensor_s;
-        } else {
-          out << "<Tensor>";
-        }
-        break;
-      }
-    case AttributeKind::ts:
-      out << "[<Tensors>]";
-      break;
-    case AttributeKind::g:
-      out << "<Graph>";
-      break;
-    case AttributeKind::gs:
-      out << "[<Graphs>]";
-      break;
-  }
-}
 
 void printAttributes(std::ostream & out, const Node * n, bool ignore_subgraph=false) {
   out << "[";
@@ -170,7 +89,8 @@ void printAttributes(std::ostream & out, const Node * n, bool ignore_subgraph=fa
     // bug by printing it out.
     out << name.toUnqualString() << "=";
 
-    printAttributeValue(out, name, n);
+    // printAttributeValue(out, name, n);
+    n->printValue(out, name);
   }
   out << "]";
 }
@@ -256,350 +176,13 @@ std::ostream& operator<<(std::ostream & out, const Graph & g) {
   return out;
 }
 
-class PrettyPrintPass {
-  Graph& graph_;
-
-  // When printing a name if there is a conflict with an existing name in the
-  // graph, record the value -> new generated name mapping
-  std::unordered_map<const Value*, const Value*> aliases_;
-
-  // The Graph already tracks unique_names_, this is just for additional ones
-  // generated during printing
-  std::unordered_map<std::string, const Value*> generated_names_;
-
-  // Cache of value names
-  std::unordered_map<const Value*, std::string> value_names_;
-
-  template<class T>
-  void zipWith(
-    at::ArrayRef<T> list_a,
-    at::ArrayRef<T> list_b,
-    std::function<void(T, T)> action
-  ) const {
-    auto it_a = list_a.begin();
-    auto it_b = list_b.begin();
-
-    if (list_a.size() != list_b.size()) {
-      AT_ERROR("Pretty printer expected 2 lists of same size");
-    }
-
-    for (; it_a != list_a.end(); ++it_a, ++it_b) {
-      action(*it_a, *it_b);
-    }
-  }
-
-  std::ostream& printValueList(
-    std::ostream & out,
-    at::ArrayRef<const Value*> list
-  ) {
-    out << "(";
-    auto delimiter = "";
-    for (const auto* value : list) {
-      out << delimiter;
-      printValue(out, value);
-      delimiter = ", ";
-    }
-    out << ")";
-    return out;
-  }
-
-  void printAssignment(
-    std::ostream& out,
-    const Value* lhs,
-    const Value* rhs,
-    const size_t level
-  ) {
-    indent(out, level);
-    printValue(out, lhs);
-    out << " = ";
-    printValue(out, rhs);
-    out << "\n";
-  }
-
-  std::ostream& printIf(
-    std::ostream & out,
-    const Node* node,
-    const size_t level
-  ) {
-    indent(out, level);
-    out << "if ";
-    const auto if_block = node->blocks()[0];
-    const auto else_block = node->blocks()[1];
-    printValue(out, node->inputs()[0]);
-    out << ":" << "\n";
-
-    // Print node contents
-    printBlock(out, if_block, level + 1);
-
-    // Print if block output
-    zipWith<const Value*>(
-      node->outputs(),
-      if_block->outputs(),
-      [&](const Value* node_output, const Value* return_input) {
-        printAssignment(out, node_output, return_input, level + 1);
-      }
-    );
-
-    indent(out, level);
-    out << "else:\n";
-    printBlock(out, else_block, level + 1);
-    zipWith<const Value*>(
-      node->outputs(),
-      else_block->outputs(),
-      [&](const Value* node_output, const Value* return_input) {
-        printAssignment(out, node_output, return_input, level + 1);
-      }
-    );
-
-    return out;
-  }
-
-  bool isValueUsedLater(
-    const Value* val
-  ) const {
-    if (aliases_.find(val) != aliases_.end()) {
-      return true;
-    }
-    return val->uses().size() > 0;
-  }
-
-  std::ostream& printLoop(
-    std::ostream & out,
-    const Node* node,
-    const size_t level
-  ) {
-    const auto body_block = node->blocks()[0];
-    aliases_[body_block->inputs()[0]] = body_block->inputs()[0];
-
-    // Add temporaries for loop-carried dependencies
-    zipWith<const Value*>(
-      node->inputs().slice(1),
-      body_block->inputs(),
-      [&](const Value* node_input, const Value* param_output) {
-        if (isValueUsedLater(param_output)) {
-          printAssignment(out, param_output, node_input, level);
-        }
-      }
-    );
-
-    zipWith<const Value*>(
-      node->outputs(),
-      node->inputs().slice(2),
-      [&](const Value* node_output, const Value* return_input) {
-        if (isValueUsedLater(node_output)) {
-          printAssignment(out, node_output, return_input, level);
-        }
-      }
-    );
-
-    // Loop header
-    indent(out, level);
-    out << "while ";
-    printValue(out, body_block->inputs()[0]);
-    out << ":\n";
-
-    // Loop body
-    printBlock(out, body_block, level + 1);
-
-    // Re-assign block outputs to inputs for next iteration
-    zipWith<const Value*>(
-      body_block->inputs(),
-      body_block->outputs(),
-      [&](const Value* param_output, const Value* return_input) {
-        if (isValueUsedLater(param_output)) {
-          printAssignment(out, param_output, return_input, level + 1);
-        }
-      }
-    );
-
-    // The block outputs are not live after the end of the block, so we can use
-    // them as the value names and avoid printing assignments
-    zipWith<const Value*>(
-      node->outputs(),
-      body_block->outputs().slice(1),
-      [&](const Value* node_output, const Value* return_input) {
-        if (isValueUsedLater(node_output)) {
-          printAssignment(out, node_output, return_input, level + 1);
-        }
-      }
-    );
-
-    return out;
-  }
-
-  std::ostream& printNode(
-    std::ostream & out,
-    const Node* node,
-    const size_t level
-  ) {
-    // if there are subblocks on this node, visit them
-    switch (node->kind()) {
-    case prim::Return:
-      // Handled elsewhere, do nothing
-      break;
-    case prim::Constant:
-      break;
-    case prim::Loop:
-      printLoop(out, node, level);
-      break;
-    case prim::If:
-      printIf(out, node, level);
-      break;
-    default:
-      indent(out, level);
-      // Print outputs
-      if (node->outputs().size() > 0) {
-        auto delim = "";
-        for (const auto* output_value : node->outputs()) {
-          out << delim;
-          printValue(out, output_value);
-          delim = ", ";
-        }
-        out << " = ";
-      }
-
-      out << node->kind().toQualString();
-
-      // Print instruction parameters
-      printValueList(out, node->inputs());
-
-      out << "\n";
-    }
-
-    return out;
-  }
-
-  std::ostream& printReturn(
-    std::ostream & out,
-    const Node* node,
-    const size_t level
-  ) {
-    indent(out, level);
-    const auto& returns = node->inputs();
-    if (returns.size() > 0) {
-      out << "return ";
-      std::string delimiter = "";
-      if (returns.size() > 1) {
-        printValueList(out, returns);
-      } else {
-        printValue(out, returns[0]);
-      }
-      out << "\n";
-    }
-    return out;
-  }
-
-  std::ostream& printBlock(
-    std::ostream & out,
-    const Block* root,
-    const size_t level
-  ) {
-    for (const auto* node : root->nodes()) {
-      printNode(out, node, level);
-    }
-
-    printNode(out, root->return_node(), level);
-
-    return out;
-  }
-
-  inline bool isNameUnique(std::string& name, const Value* val) const {
-    auto generated_name_value = generated_names_.find(name);
-    if (generated_name_value != generated_names_.end() &&
-        generated_name_value->second != val) {
-      // Found a generated name match, check that it's for a different value
-      return false;
-    }
-    return graph_.uniqueNames().find(name) == graph_.uniqueNames().end();
-  }
-
-  std::ostream& printValue(
-    std::ostream & out,
-    const Value* val
-  ) {
-    auto cached_name = value_names_.find(val);
-    if (cached_name != value_names_.end()) {
-      // If this value has been seen before, print out cached name
-      out << cached_name->second;
-      return out;
-    }
-
-    const auto node = val->node();
-    if (node->kind() == prim::Constant) {
-      printAttributeValue(out, node->attributeNames()[0], node);
-      return out;
-    }
-
-    auto name_source = val;
-
-    auto aliased_name = aliases_.find(val);
-    if (aliased_name != aliases_.end()) {
-      name_source = aliased_name->second;
-    }
-
-    auto name = name_source->uniqueName();
-
-    bool using_generated_name = false;
-    if (isdigit(name.at(0))) {
-      std::stringstream ss;
-      ss << "t" << name;
-      name = ss.str();
-      using_generated_name = true;
-    } else if (name.find_last_of('.') != std::string::npos) {
-      // Make unique name a valid variable name (e.g. a.1 -> a1)
-      name.erase(std::remove(name.begin(), name.end(), '.'), name.end());
-      using_generated_name = true;
-    }
-
-    if (using_generated_name) {
-      // Make sure name is unique
-      size_t suffix = 0;
-      while (!isNameUnique(name, name_source)) {
-        std::stringstream ss;
-        ss << name << suffix;
-        name = ss.str();
-        ++suffix;
-      }
-
-      // These names aren't in the Graph's list of names but we still need to
-      // make sure there are no name conflicts
-      generated_names_[name] = name_source;
-    }
-
-    value_names_[val] = name;
-    out << name;
-    return out;
-  }
-
-public:
-  PrettyPrintPass(Graph& graph) : graph_(graph) {}
-
-  std::ostream& run(std::ostream & out) {
-    out << "def script";
-    const Node* params = graph_.block()->param_node();
-    printValueList(out, params->outputs());
-    out << ":\n";
-
-    // Print body
-    printBlock(out, graph_.block(), 1);
-
-    printReturn(out, graph_.block()->return_node(), 1);
-
-    return out;
-  }
-};
-
 std::ostream& Graph::prettyPrint(std::ostream & out) {
-  PrettyPrintPass pass(*this);
-
-  pass.run(out);
-
+  PrettyPrint(out, *this);
   return out;
 }
 
 void Graph::dumpPretty() {
-  PrettyPrintPass pass(*this);
-  pass.run(std::cout);
+  PrettyPrint(std::cout, *this);
 }
 
 static void checkSameDevice(const Node* node) {
